slider/models.py
from django.db import models
from filer.fields.image import FilerImageField


class SliderItem(models.Model):
    title = models.CharField('Название', max_length=255)
    description = models.TextField('Описание', blank=True)
    image = FilerImageField(
        verbose_name='Изображение',
        on_delete=models.SET_NULL,
        null=True,
        blank=True,
        related_name='slider_items',
    ) # type: ignore
    order = models.PositiveIntegerField('Порядок', default=0, db_index=True)
    is_active = models.BooleanField('Активен', default=True)

    class Meta:
        verbose_name = 'Слайд'
        verbose_name_plural = 'Слайды'
        ordering = ['order']

    def __str__(self):
        return self.title
