pyproject.toml
[project]
name = "django-slider-project-3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django==5.2",
    "django-admin-sortable2==2.1.10",
    "django-filer==3.2.2",
    "easy-thumbnails==2.10",
    "pillow==10.3.0",
    "pymysql==1.1.1",
]
